requirements.txt
scapy>=2.5.0
scikit-learn>=1.3.0
flask>=2.0.0
flask-socketio>=5.0.0
requests>=2.25.0
sqlalchemy>=2.0.0
pyyaml>=6.0
python-dotenv>=1.0.0
# Add any other dependencies you need